move image loading to separated file
