+++ src/js/components/image-loading.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Loading() {
+    return <div className='image-overlay'>
+               <div className='three-bounce spinner'>
+                   <div className='bounce1'></div>
+                   <div className='bounce2'></div>
+                   <div className='bounce3'></div>
+               </div>
+           </div>;
+}
